add netlify function
